Revert "Fix default icon shown for BGMApp bundle."
This reverts commit 0d6a0c847044adda8dbf99157202c5f6e910e5af.

I've tested the icon bug on a couple of other machines now and haven't been
able to reproduce it. My "fix" did cause the bug on those machines, though,
despite fixing it on mine. Still not sure exactly what's going on, but it seems
to be a problem with my development environment.