Fix another crash when BGMDevice's volume is changed.
Fixes the same issue as dfca337b63e2c7d50e1e674a3697c7e02e05939a, but in
BGMOutputVolumeMenuItem instead of BGMStatusBarItem.

I think the problem was that it captured a weak reference in a C++
lambda, but it would capture by (C++) reference and when the reference
was used it would be referencing invalid memory. The fix to have the
lambda capture by value instead.

See #202.